refactor: FSD Phase 6b — 学習系の単一画面フックを pages の model へ移す (FRESTYLE-167)
単一画面からのみ使われ store にも他 hook にも依存しない学習系 hook 9 個を、
その画面と同じ Slice の model へ同居させる。

- useCourses → pages/courses/model
- useTeachingMaterials / useLessonProgress / useChapterResume / useNextCourse
  → pages/course-detail/model
- useExerciseList → pages/exercises/model
- useExerciseDetail → pages/exercise-detail/model
- useExerciseLanguageSummary → pages/exercise-languages/model
- useLearningReport → pages/learning-report/model

依存を調べたところ FreStyle の hook はほぼ 1 画面 = 1 hook で、複数画面で
共有される横断機能は少ない(認証と AI チャットくらい)。単一画面専用の hook は
その画面の model に置くのが最も見つけやすく FSD 的にも正しい(features は
再利用単位で、1 画面専用は page の model)。barrel は作らない(同一 Slice
からしか使わないため深い相対参照でよい)。

useTeachingMaterialEditor は useNoteEditor(notes と course-detail で共有)に
依存するため対象外。共有 hook の配置を決める後続 PR で扱う。

page ui からは ../model/useX、model 内テストからは ../useX の相対参照に統一。
pages 間 import が無いことを確認済み(境界 lint 緑)。

既存テストの assert は変更していない。テスト件数 1184 で不変。
