+++ frontend/src/pages/course-detail/model/__tests__/useLessonProgress.test.ts
