+++ frontend/src/pages/course-detail/model/__tests__/useNextCourse.test.ts
