+++ frontend/src/pages/course-detail/model/__tests__/useTeachingMaterials.test.ts
