+++ frontend/src/pages/course-detail/model/useLessonProgress.ts
