+++ frontend/src/pages/course-detail/model/useNextCourse.ts
