+++ frontend/src/pages/course-detail/model/useTeachingMaterials.ts
