+++ frontend/src/pages/course-detail/ui/CourseDetailPage.tsx
@@ -18,11 +18,11 @@ import ConfirmModal from '@/shared/ui/ConfirmModal';
 import Loading from '@/shared/ui/Loading';
 import NoteMarkdownEditor from '@/components/NoteMarkdownEditor';
 import MarkdownTableOfContents from '@/shared/ui/MarkdownTableOfContents';
-import { useTeachingMaterials } from '@/hooks/useTeachingMaterials';
+import { useTeachingMaterials } from '../model/useTeachingMaterials';
 import { useTeachingMaterialEditor } from '@/hooks/useTeachingMaterialEditor';
-import { useChapterResume } from '@/hooks/useChapterResume';
-import { useNextCourse } from '@/hooks/useNextCourse';
-import { useLessonProgress } from '@/hooks/useLessonProgress';
+import { useChapterResume } from '../model/useChapterResume';
+import { useNextCourse } from '../model/useNextCourse';
+import { useLessonProgress } from '../model/useLessonProgress';
 import { useMobilePanelState } from '@/shared/lib/hooks/useMobilePanelState';
 import { useLocalStorage } from '@/shared/lib/hooks/useLocalStorage';
 import { useToast } from '@/shared/lib/hooks/useToast';
