+++ frontend/src/pages/courses/model/useCourses.ts
