+++ frontend/src/pages/exercise-detail/model/__tests__/useExerciseDetail.test.ts
