+++ frontend/src/pages/exercise-detail/model/useExerciseDetail.ts
