+++ frontend/src/pages/learning-report/__tests__/LearningReportPage.test.tsx
@@ -1,10 +1,10 @@
 import { describe, it, expect, vi } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
 import LearningReportPage from '../ui/LearningReportPage';
-import * as useLearningReportModule from '@/hooks/useLearningReport';
+import * as useLearningReportModule from '../model/useLearningReport';
 import type { LearningReport } from '@/entities/learning-report';
 
-vi.mock('@/hooks/useLearningReport');
+vi.mock('../model/useLearningReport');
 
 const mockReport: LearningReport = {
   id: 1,
